requirements.txt
email-validator
Jinja2
pydantic
python-decouple
pytz
requests
